agents/pyproject.toml
[project]
name = "foundry-hosted-langchain-agents"
version = "0.1.0"
description = "Langchain hosted agent for Microsoft Foundry"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-agentserver-responses==1.0.0b5",
    "azure-ai-agentserver-core==2.0.0b3",
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.7",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-azure-ai[opentelemetry]>=1.2.3",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = ["vendor"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
